+++ tests/integration/base.py
@@ -161,7 +161,7 @@ class DBTIntegrationTest(unittest.TestCase):
                     'thrift': {
                         'type': 'iomete',
                         'host': os.getenv('DBT_IOMETE_HOST_NAME'),
-                        'cluster': os.getenv('DBT_IOMETE_CLUSTER_NAME'),
+                        'lakehouse': os.getenv('DBT_IOMETE_LAKEHOUSE'),
                         'account_number': os.getenv('DBT_IOMETE_ACCOUNT_NUMBER'),
                         'user': os.getenv('DBT_IOMETE_USER_NAME'),
                         'password': os.getenv('DBT_IOMETE_PASSWORD'),